Reorganize settings into multiple pages. The main page has icons, titles and summaries. Fixes #960.
