+++ src/main/java/de/dennisguse/opentracks/settings/OpenTracksSettingsFragment.java
@@ -1,0 +1,24 @@
+package de.dennisguse.opentracks.settings;
+
+import android.os.Bundle;
+import android.util.Log;
+
+import androidx.fragment.app.DialogFragment;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+
+import de.dennisguse.opentracks.R;
+
+public class OpenTracksSettingsFragment extends PreferenceFragmentCompat {
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        addPreferencesFromResource(R.xml.settings_open_tracks);
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        ((SettingsActivity) getActivity()).getSupportActionBar().setTitle(R.string.settings_open_tracks_title);
+    }
+}
