+++ src/main/java/de/dennisguse/opentracks/settings/UserInterfaceSettingsFragment.java
@@ -1,0 +1,67 @@
+package de.dennisguse.opentracks.settings;
+
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.os.Bundle;
+
+import androidx.fragment.app.DialogFragment;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+
+import de.dennisguse.opentracks.R;
+
+public class UserInterfaceSettingsFragment extends PreferenceFragmentCompat {
+
+    private final SharedPreferences.OnSharedPreferenceChangeListener sharedPreferenceChangeListener = (sharedPreferences, key) -> {
+        if (PreferencesUtils.isKey(R.string.night_mode_key, key)) {
+            getActivity().runOnUiThread(PreferencesUtils::applyNightMode);
+        }
+    };
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        addPreferencesFromResource(R.xml.settings_user_interface);
+
+        Preference customLayoutPreference = findPreference(getString(R.string.stats_custom_layout_key));
+        customLayoutPreference.setOnPreferenceClickListener((preference) -> {
+            Intent intent = new Intent(getContext(), SettingsCustomLayoutActivity.class);
+            startActivity(intent);
+            return true;
+        });
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        ((SettingsActivity) getActivity()).getSupportActionBar().setTitle(R.string.settings_ui_title);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        PreferencesUtils.registerOnSharedPreferenceChangeListener(sharedPreferenceChangeListener);
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        PreferencesUtils.unregisterOnSharedPreferenceChangeListener(sharedPreferenceChangeListener);
+    }
+
+    @Override
+    public void onDisplayPreferenceDialog(Preference preference) {
+        DialogFragment dialogFragment = null;
+
+        if (preference instanceof ResetDialogPreference) {
+            dialogFragment = ResetDialogPreference.ResetPreferenceDialog.newInstance(preference.getKey());
+        }
+
+        if (dialogFragment != null) {
+            dialogFragment.setTargetFragment(this, 0);
+            dialogFragment.show(getParentFragmentManager(), getClass().getSimpleName());
+            return;
+        }
+
+        super.onDisplayPreferenceDialog(preference);
+    }
+}
