+++ src/main/res/values/settings.xml
@@ -10,7 +10,41 @@
     <string name="stats_fullscreen_while_recording_key" translatable="false">trackdetail_fullscreen_while_recording</string>
     <bool name="stats_fullscreen_while_recording_default" translatable="false">false</bool>
 
+    <!-- Main settings -->
+    <string name="settings_open_tracks_key" translatable="false">settingsOpenTracks</string>
+    <string name="settings_open_tracks_title" translatable="false">Open Tracks</string>
+    <string name="settings_open_tracks_summary">All About OpenTracks Project</string>
+
+    <string name="settings_defaults_key" translatable="false">settingsDefaults</string>
+    <string name="settings_defaults_title">Default Units and Activity</string>
+    <string name="settings_defaults_summary">Preferred Units, Default Activity, Track Default Name</string>
+
+    <string name="settings_ui_key" translatable="false">settingsUI</string>
+    <string name="settings_ui_title">User Interface</string>
+    <string name="settings_ui_summary">Appearance, Recording Layout, Theme</string>
+
+    <string name="settings_gps_key" translatable="false">settingsGps</string>
+    <string name="settings_gps_title">GPS</string>
+    <string name="settings_gps_summary">Time interval, Distance interval, Max Distance, Accuracy</string>
+
+    <string name="settings_sensors_key" translatable="false">settingsSensorsGps</string>
+    <string name="settings_sensors_title">Bluetooth Sensors</string>
+    <string name="settings_sensors_summary">Cycling Sensors, Wheel Circumference, Running Sensors</string>
+
+    <string name="settings_announcements_key" translatable="false">settingsAnnouncements</string>
+    <string name="settings_announcements_title">Voice Announcements</string>
+    <string name="settings_announcements_summary">Time, Distance, Voice Speed</string>
+
+    <string name="settings_import_export_key" translatable="false">settingsImportExport</string>
+    <string name="settings_import_export_title">Import/Export</string>
+    <string name="settings_import_export_summary">Import, Export, Auto Export</string>
+
+    <!-- Setting's categories -->
+    <string name="settings_units_category">Units</string>
+    <string name="settings_activity_defaults_category">Activity\'s Defaults</string>
+
     <string name="settings_reset_key" translatable="false">settingsReset</string>
+    <string name="settings_reset_summary">Reset Settings to Default Values</string>
     <string name="settings_layout_reset_key" translatable="false">settingsLayoutReset</string>
 
     <string name="settings_sensor_bluetooth_heart_rate_key" translatable="false">bluetoothSensor</string>
