+++ src/main/res/xml/settings_announcements.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:title="@string/settings_announcements_title">
+
+    <ListPreference
+        android:defaultValue="@string/voice_announcement_frequency_default"
+        android:entryValues="@array/voice_announcement_frequency_values"
+        android:key="@string/voice_announcement_frequency_key"
+        android:title="@string/settings_announcements_totaltime_frequency"
+        app:useSimpleSummaryProvider="true" />
+
+    <ListPreference
+        android:defaultValue="@string/voice_announcement_distance_default"
+        android:entryValues="@array/voice_announcement_distance_values"
+        android:key="@string/voice_announcement_distance_key"
+        android:title="@string/settings_announcements_distance_frequency"
+        app:useSimpleSummaryProvider="true" />
+
+    <ListPreference
+        android:defaultValue="@string/voice_speed_rate_default"
+        android:entryValues="@array/voice_speed_rate_values"
+        android:entries="@array/voice_speed_rate_values"
+        android:key="@string/voice_speed_rate_key"
+        android:title="@string/menu_voice_rate"
+        app:useSimpleSummaryProvider="true" />
+
+</PreferenceScreen>
