+++ src/main/res/xml/settings_open_tracks.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    android:title="@string/settings_open_tracks_title">
+
+    <Preference android:title="@string/menu_help">
+        <intent
+            android:targetClass="de.dennisguse.opentracks.HelpActivity"
+            android:targetPackage="@string/applicationId" />
+    </Preference>
+    <Preference android:title="@string/about_preference_title">
+        <intent
+            android:targetClass="de.dennisguse.opentracks.AboutActivity"
+            android:targetPackage="@string/applicationId" />
+    </Preference>
+
+</PreferenceScreen>
