+++ src/main/res/xml/settings_sensors.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:title="@string/settings_sensors_title">
+
+    <PreferenceCategory android:title="@string/settings_generic_sensor">
+        <de.dennisguse.opentracks.settings.bluetooth.BluetoothLeHeartRatePreference
+            android:defaultValue="@string/sensor_type_value_none"
+            android:key="@string/settings_sensor_bluetooth_heart_rate_key"
+            android:title="@string/sensor_state_heart_rate" />
+    </PreferenceCategory>
+
+    <PreferenceCategory android:title="@string/settings_cycling_sensor">
+        <de.dennisguse.opentracks.settings.bluetooth.BluetoothLeCyclingCadenceAndSpeedPreference
+            android:defaultValue="@string/sensor_type_value_none"
+            android:key="@string/settings_sensor_bluetooth_cycling_cadence_key"
+            android:title="@string/settings_sensor_cycling_cadence" />
+        <de.dennisguse.opentracks.settings.bluetooth.BluetoothLeCyclingPowerPreference
+            android:defaultValue="@string/sensor_type_value_none"
+            android:key="@string/settings_sensor_bluetooth_cycling_power_key"
+            android:title="@string/sensor_state_power" />
+        <de.dennisguse.opentracks.settings.bluetooth.BluetoothLeCyclingCadenceAndSpeedPreference
+            android:defaultValue="@string/sensor_type_value_none"
+            android:key="@string/settings_sensor_bluetooth_cycling_speed_key"
+            android:title="@string/settings_sensor_cycling_distance_speed" />
+        <EditTextPreference
+            android:defaultValue="@string/settings_sensor_bluetooth_cycling_speed_wheel_circumference_default"
+            android:inputType="number"
+            android:key="@string/settings_sensor_bluetooth_cycling_speed_wheel_circumference_key"
+            android:title="@string/settings_sensor_wheel_circumference"
+            app:useSimpleSummaryProvider="true" />
+    </PreferenceCategory>
+
+    <PreferenceCategory android:title="@string/settings_running_sensor">
+        <de.dennisguse.opentracks.settings.bluetooth.BluetoothLeRunningSpeedAndCadencePreference
+            android:defaultValue="@string/sensor_type_value_none"
+            android:key="@string/settings_sensor_bluetooth_running_speed_and_cadence_key"
+            android:title="@string/settings_sensor_running_speed_and_cadence" />
+    </PreferenceCategory>
+
+</PreferenceScreen>
